refactor(types): reorganize and enhance type definitions across modules
- Introduced a new `worker-types.ts` file to centralize worker-related type definitions, improving clarity and maintainability.
- Refactored imports across various modules to utilize the new `worker-types`, ensuring consistency in type usage.
- Moved `ChessPiece` and `MoveCoords` interfaces to appropriate files, enhancing type organization and reducing redundancy.
- Updated `AGENTS.md` to reflect changes in type organization and clarify the purpose of key directories.
- Enhanced documentation for type definitions, improving usability and understanding for developers.
