+++ src/index.ts
@@ -4,6 +4,6 @@ export { getAnalyzeError, isReplayError } from '#core/analyze-errors';
 
 // Types
 export type { AnalyzeOptions, AnalyzeResult } from '#types/analysis';
-export type { HeatmapData } from '#types/tracker';
+export type { HeatmapData } from '#trackers/heatmap-types';
 export type { AnalyzeError, ReplayError, ReplayErrorReason } from '#types/errors';
 export type { ReplayMode } from '#replay/replay-mode';
