+++ src/pgn/parse-options.ts
@@ -1,4 +1,10 @@
-import type { ParsePgnOptions, StandaloneParseOptions } from '#types/parse-pgn';
+import type { ParsePgnOptions } from '#types/parse-pgn';
+
+/** Resolved standalone-parse options shared by {@link parsePGN} and {@link streamParsePGN}. */
+export interface StandaloneParseOptions {
+    parseHeaders: boolean;
+    maxGames: number;
+}
 
 /**
  * Resolve {@link ParsePgnOptions} for the standalone parse entry points.
